requirements.txt
# Core runtime
pydantic>=2.0.0
python-dotenv>=1.0.0
requests>=2.31.0
PyYAML>=6.0
tqdm>=4.0.0
tiktoken>=0.7.0
numpy>=1.24.0
pandas>=2.2.0
matplotlib>=3.8.0
openpyxl>=3.1.0

# Chunking / retrieval / storage
chonkie[all]
cohere>=5.14.2
FlagEmbedding>=1.2.0
qdrant-client>=1.9.0
aiosqlite>=0.20.0
httpx[socks]>=0.24.0
socksio>=1.0.0

# APIs / servers
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
openai>=1.0.0
google-genai>=1.10.0
google-generativeai>=0.8.0

# Evaluation
rouge-score>=0.1.2
nltk>=3.8.0
sentence-transformers>=2.2.0

# File processing
pypdf>=3.0.0
pypdfium2>=4.30.0
Pillow>=10.0.0
